Refatora autenticação, dependências e RPA conforme issue #69
- Move arquivos de autenticação (security.py, dependencies.py, utils.py) para src/routes/auth/
- Renomeia _get_user_uuid para get_current_user_uuid e ajusta todos os imports
- Atualiza referências aos helpers, centralizando utilitários realmente genéricos em src/utils
- Ajusta imports de mensagens para src.siscan.messages
- Cria src/siscan/rpa/runner.py para centralizar execução do RPA
- Atualiza main.py e rotas para refletir a nova organização

Refs #69
