+++ src/routes/auth/utils.py
@@ -3,10 +3,9 @@ from cryptography.hazmat.primitives import hashes
 from cryptography.hazmat.primitives.asymmetric import padding
 from datetime import datetime, timedelta
 from fastapi.security import OAuth2PasswordBearer
-from playwright.async_api import async_playwright
 from typing import Any, Dict, Optional
 
-from src.env import PRODUCTION, TAKE_SCREENSHOT, private_key, public_key
+from src.env import private_key, public_key
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="security/token")
 
@@ -57,37 +56,3 @@ def decode_access_token(token: str) -> Dict[str, Any]:
     """Decode a JWT token using the RSA public key."""
     return jwt.decode(token, public_key, algorithms=["RS256"])
 
-
-async def run_rpa(form_type, data):
-    """Executa o fluxo do RPA utilizando Playwright assíncrono."""
-    screenshots = []
-    async with async_playwright() as p:
-        browser = await p.chromium.launch(headless=True)
-        page = await browser.new_page()
-        # TODO: implementar login no SISCAN usando CPF/senha de users db
-
-        # TODO: navegar até o formulário e preencher campos com 'data'
-        # Exemplo: page.handle_fill("input[name=campo1]",
-        """
-        req = RequisicaoExameMamografiaRastreio(
-            base_url=SISCAN_URL, user=SISCAN_USER, password=SISCAN_PASSWORD
-        )
-
-        req._context = SiscanBrowserContext(headless=headless)
-
-        await req.authenticate()
-        await req.preencher(json_data)
-        """
-        # informations = req.context.information_messages
-        if not PRODUCTION and TAKE_SCREENSHOT:
-            for i in range(1, 4):
-                path = f"static/tmp/{form_type}_step{i}.png"
-                await page.screenshot(path=path)
-                screenshots.append(path)
-
-        if PRODUCTION:
-            await page.click("button[type=submit]")
-
-        await browser.close()
-
-    return {"success": True, "screenshots": screenshots}
