+++ src/routes/user.py
@@ -3,11 +3,11 @@ from sqlalchemy.exc import IntegrityError
 
 from src.env import get_db
 from src.models import User
+from src.routes.auth.dependencies import api_key_required
+from src.routes.auth.utils import encrypt_password, oauth2_scheme, \
+    decode_access_token
 from src.siscan.schema.auth_schema import LoginInput
-from src.utils import messages as msg
-from src.utils.dependencies import api_key_required
-from src.utils.helpers import encrypt_password, decode_access_token, \
-    oauth2_scheme
+from src.siscan import messages as msg
 
 router = APIRouter(prefix="/user", tags=["user"])
